Implemented threading strategy to prevent UI updates outside the main queue using the Decorator pattern. This way, the UI is agnostic of threading logic/details.
Threading is dealt with in the Composition layer.
