+++ FeedAppiOSTest/Feed UI/Controllers/FeedUIIntegrationTests.swift
@@ -288,6 +288,33 @@ final class FeedUIIntegrationTests: XCTestCase {
         
         XCTAssertEqual(newView.renderedImage, imageData)
     }
+    
+    func test_loadFeedCompletion_dispatchesFromBackgroundToMainThread() {
+            let (sut, loader) = makeSUT()
+            sut.loadViewIfNeeded()
+
+            let exp = expectation(description: "Wait for background queue")
+            DispatchQueue.global().async {
+                loader.completeFeedLoading(at: 0)
+                exp.fulfill()
+            }
+            wait(for: [exp], timeout: 1.0)
+        }
+
+        func test_loadImageDataCompletion_dispatchesFromBackgroundToMainThread() {
+            let (sut, loader) = makeSUT()
+
+            sut.loadViewIfNeeded()
+            loader.completeFeedLoading(with: [makeImage()])
+            _ = sut.simulateFeedImageViewVisible(at: 0)
+
+            let exp = expectation(description: "Wait for background queue")
+            DispatchQueue.global().async {
+                loader.completeImageLoading(with: self.anyImageData(), at: 0)
+                exp.fulfill()
+            }
+            wait(for: [exp], timeout: 1.0)
+        }
 
     // MARK: - Helpers
